Separated Player Script
Separated the Player Script into several different scripts. Player is a base script, Movement handles movement controls, and PlayerAnimation will handle the visuals of the character.
Also separated the visual of the character from the functioning logic portion.
Changed the ladder prefab to have a much thinner collider, so the player will need to be more centered on it to use it.