Attempt at fixing the falling bug
Created new platform prefabs and variants so the transform.position will be at the lowest point in the world to ensure player will not ignore it when it is on top of it.
That didn't work. No clue why it's still falling through every platform other than the bottom one 🤷‍♂️