Fixed Falling Bug - Added Ladder Script
Fixed falling bug by changing all movement to be kinematic. All movement is moved through code now, and if the player is deemed to be grounded, a raycast will be cast and the player will be placed on top of the platform they're grounded on.

Added ladder script to make adding more more streamlined. They will automatically find the platform they're on, and the one directly above them and create their bounding box to allow the player to climb up and down them.