Implemented Barrel Base Movement
Made animations for DK and made him be the barrel spawner. It now looks like he is throwing the barrels. I had to move the girl and some platforms up so there was room for the barrels.

Changed some of the code to allow a more consistent and predictable movement path.

Placed the fire barrel at the beginning of the level to create a fire enemy, or to kill the player if they touch the fire. Created the animation for it as well.

Changed the character sprites to face right and changed the code to keep them facing the correct direction.